Add path-planning-specific stats
Added batch stats for manhattan distance and number of times the goal
was reached, and run stats for manhattan distance from goal.

+++ AdversarialCoverageSimulator/src/adsim/stats/SampledVariableLong.java
@@ -41,11 +41,16 @@ public class SampledVariableLong extends SampledVariable {
 	}
 
 
+	/**
+	 * Gets the mean of the samples
+	 * 
+	 * @return the mean
+	 */
 	public double mean() {
 		if (this.nSamples < 1) {
 			return Double.NaN;
 		}
-		return (this.runningSum / this.nSamples);
+		return (((double) this.runningSum) / ((double) this.nSamples));
 	}
